Slimdown pass: docs helpers, pack run, guardrail stages
