--- scripts/codex-devtools.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-set -euo pipefail
-
-exec codex -c 'mcp_servers.chrome-devtools.enabled=true' "$@"
