+++ scripts/pack-audit.mjs
@@ -1,12 +1,9 @@
 #!/usr/bin/env node
 
-import { execFile } from 'node:child_process';
 import { rm } from 'node:fs/promises';
 import path from 'node:path';
 import process from 'node:process';
-import { promisify } from 'node:util';
-
-const execFileAsync = promisify(execFile);
+import { runPack } from './lib/npm-pack.js';
 
 const REQUIRED_FILES = [
   'dist/bin/codex-orchestrator.js',
@@ -47,27 +44,6 @@ function parseByteLimit(value) {
   const parsed = Number(value);
   if (!Number.isFinite(parsed) || parsed < 0) return null;
   return Math.floor(parsed);
-}
-
-async function runPack() {
-  const { stdout } = await execFileAsync(
-    'npm',
-    ['pack', '--json', '--ignore-scripts'],
-    {
-      env: { ...process.env, npm_config_ignore_scripts: 'true' },
-      maxBuffer: 10 * 1024 * 1024
-    }
-  );
-  const trimmed = String(stdout ?? '').trim();
-  if (!trimmed) {
-    throw new Error('npm pack produced no output');
-  }
-  const parsed = JSON.parse(trimmed);
-  const record = Array.isArray(parsed) ? parsed[0] : parsed;
-  if (!record || typeof record !== 'object') {
-    throw new Error('npm pack output did not include a record');
-  }
-  return record;
 }
 
 function collectFilePaths(record) {
