+++ scripts/pack-smoke.mjs
@@ -1,34 +1,12 @@
 #!/usr/bin/env node
 
-import { execFile, spawn } from 'node:child_process';
+import { spawn } from 'node:child_process';
 import { mkdtemp, rm, writeFile } from 'node:fs/promises';
 import os from 'node:os';
 import path from 'node:path';
 import process from 'node:process';
-import { promisify } from 'node:util';
+import { runPack } from './lib/npm-pack.js';
 
-const execFileAsync = promisify(execFile);
-
-async function runPack() {
-  const { stdout } = await execFileAsync(
-    'npm',
-    ['pack', '--json', '--ignore-scripts'],
-    {
-      env: { ...process.env, npm_config_ignore_scripts: 'true' },
-      maxBuffer: 10 * 1024 * 1024
-    }
-  );
-  const trimmed = String(stdout ?? '').trim();
-  if (!trimmed) {
-    throw new Error('npm pack produced no output');
-  }
-  const parsed = JSON.parse(trimmed);
-  const record = Array.isArray(parsed) ? parsed[0] : parsed;
-  if (!record || typeof record !== 'object') {
-    throw new Error('npm pack output did not include a record');
-  }
-  return record;
-}
 
 async function runCommand(command, args, options = {}) {
   await new Promise((resolve, reject) => {
